feat(types): contextual integer literal typing [verify-poc]
Integer literals take the integer type their context requires: declared and
assigned types, arguments, returns, indices and slice bounds, and the typed
peer operand of arithmetic, comparison, equality, and bitwise operators in
either order. A literal that does not fit is an error at the literal instead
of a silent fallback to int; negated literals range-check their negated
value; the signedness-mix diagnostic carries a source position.

Emission is unchanged. The verification decoder compiles to identical C with
all 185 literal casts removed, pinned by a test.

+++ typechecker/typechecker_test.go
@@ -9,11 +9,11 @@ func TestTypeChecker_IntegerLiteral(t *testing.T) {
 		input    string
 		expected string
 	}{
-		{"5", "u8"},    // Small positive literal infers u8
-		{"42", "u8"},   // Small positive literal infers u8
-		{"300", "u16"}, // Larger positive literal infers u16
-		// Note: -5 is parsed as prefix expression, not negative literal
-		// So we test positive literals only for now
+		// A literal with no context is int; typed contexts are exercised by
+		// the compiler's literal inference tests.
+		{"5", "int"},
+		{"42", "int"},
+		{"300", "int"},
 	}
 
 	for _, tt := range tests {
